requirements.txt
kubernetes
pyhelm
pyyaml
